Create zip files in build-bin
