+++ automation/build-bin.ts
@@ -1,9 +1,21 @@
+import * as Promise from 'bluebird';
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as filehound from 'filehound';
+import * as archiver from 'archiver';
 import { exec as execPkg } from 'pkg';
+import * as os from 'os';
+import * as packageJSON from '../package.json';
+import * as mkdirp from 'mkdirp';
 
+const version = 'v' + packageJSON.version;
 const ROOT = path.join(__dirname, '..');
+const outputFile = path.join(
+	ROOT,
+	'build-zip',
+	`balena-cli-${version}-${os.platform()}-${os.arch()}.zip`,
+);
+const mkdirpAsync = Promise.promisify<string | null, string>(mkdirp);
 
 console.log('Building package...\n');
 
@@ -33,4 +45,32 @@ execPkg(['--target', 'host', '--output', 'build-bin/balena', 'package.json'])
 				),
 			);
 		});
+	})
+	.then(() => {
+		console.log('Publishing build...');
+		return mkdirpAsync(path.dirname(outputFile));
+	})
+	.then(() => {
+		return new Promise((resolve, reject) => {
+			console.log('Zipping build...');
+
+			let archive = archiver('zip', {
+				zlib: { level: 7 },
+			});
+			archive.directory(path.join(ROOT, 'build-bin'), 'balena-cli');
+
+			let outputStream = fs.createWriteStream(outputFile);
+
+			outputStream.on('close', resolve);
+			outputStream.on('error', reject);
+
+			archive.on('error', reject);
+			archive.on('warning', console.warn);
+
+			archive.pipe(outputStream);
+			archive.finalize();
+		});
+	})
+	.then(() => {
+		console.log('Build zipped');
 	});
